requirements.txt
pandas
numpy
matplotlib
seaborn
requests
plotly
plotly-geo
kaleido
pymongo
python-dotenv